requirements.txt
matplotlib           3.3.4
numpy                1.19.5
pandas               1.1.5
Pillow               8.4.0
PyYAML               6.0
scipy                1.5.4
scikit-learn         0.24.2
tensorboardX         2.4.1
tifffile             2020.9.3
torch                1.10.1+cu113
torchaudio           0.10.1+cu113
torchvision          0.11.2+cu113
